Add files via upload
AIM :

To create a login and signup page . At login page we have to provide the interface to users to enter in to the website . In signup we collect the users name and password and evaluate them to create his profile

SCOPE

CREATE A SIGNUP FORM:

A signup form is a web page where users enter the information required to access that website’s services. The information collected is determined by the nature of the website and the services it offers. Most signup forms require a name, email address, username, and password.Here we take name, username, and password and confirm�password are required.

CREATE A SECURE PASSWORD:

To create a secure password, password validation is required. Passwords, when carefully created and protected, increase safe and secure interactions online and in the workplace and can prevent password cracking. To maximize the strength and efficacy of passwords, organizations often establish password policies. These policies are designed to help users create strong passwords and adopt best practices for managing login credentials. Below are a few examples of the practices that contribute to effective password management and creation:

 A minimum length of eight characters and a maximum between 16 to 64 characters. While there is no limit to the length of a password, it does reach a point of diminishing returns.

 Include both uppercase and lowercase letters with case sensitivity. This increases the number of variables at play and, therefore, its difficulty.

 Use at least one number.

 Avoid using easily guessed elements such as names of children, pet names and birthdays.

CONFIRM PASSWORD VALIDATION

It checks that the password entered by the user is same as this confirm password fields. To create a valid password, both the password and confirm password fields value must be matched.

CREATE A LOGIN FORM: A Login form is used to enter authentication credentials to access a restricted page or form. The login form contains a field for the username and another for the password. When the login form is submitted its underlying code checks that the credentials are authentic, giving the user can access the restricted page. If a user is not able to provide authentic credentials they will not be able to proceed past the login form.

PASSWORD AUTHENTICATION :

Password authentication is a process that involves a user inputting a unique ID and key that are then checked against stored credentials. You’ve likely been asked to create and/or enter a password to gain access to a personal account, whether that’s a social media platform or an online banking tool. When you do so, you’re taking part in a password authentication system that keeps your sensitive information safe from unauthorized users. The idea is that only the person who created and set the password will know it and be able to repeat it, thus verifying their identity as the original user and providing access to restricted data and resources. In password authentication, the user must supply a password for each server, and the administrator must keep track of the name and password for each user, typically on separate servers